refactor: 서비스 패키지 웹-도메인 분리
