+++ src/main/java/com/devnity/devnity/domain/admin/scheduler/AdminScheduler.java
@@ -1,6 +1,6 @@
 package com.devnity.devnity.domain.admin.scheduler;
 
-import com.devnity.devnity.domain.admin.service.AdminInvitationService;
+import com.devnity.devnity.web.admin.service.AdminInvitationService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
