+++ src/main/java/com/devnity/devnity/domain/gather/event/GatherEventHandler.java
@@ -1,6 +1,6 @@
 package com.devnity.devnity.domain.gather.event;
 
-import com.devnity.devnity.domain.gather.service.GatherUtilService;
+import com.devnity.devnity.web.gather.service.GatherUtilService;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.event.EventListener;
