+++ src/main/java/com/devnity/devnity/web/admin/controller/AdminInvitationController.java
@@ -3,7 +3,7 @@ package com.devnity.devnity.web.admin.controller;
 import com.devnity.devnity.web.common.dto.ApiResponse;
 import com.devnity.devnity.domain.admin.dto.InvitationDto;
 import com.devnity.devnity.domain.admin.dto.request.InvitationRequest;
-import com.devnity.devnity.domain.admin.service.AdminInvitationService;
+import com.devnity.devnity.web.admin.service.AdminInvitationService;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
