+++ src/test/java/com/devnity/devnity/domain/introduction/controller/IntroductionControllerTest.java
@@ -22,7 +22,7 @@ import com.devnity.devnity.domain.introduction.entity.Introduction;
 import com.devnity.devnity.domain.introduction.entity.IntroductionComment;
 import com.devnity.devnity.domain.introduction.respository.IntroductionCommentRepository;
 import com.devnity.devnity.domain.introduction.respository.IntroductionRepository;
-import com.devnity.devnity.domain.introduction.service.IntroductionService;
+import com.devnity.devnity.web.introduction.service.IntroductionService;
 import com.devnity.devnity.web.user.dto.request.SaveIntroductionRequest;
 import com.devnity.devnity.domain.user.entity.Course;
 import com.devnity.devnity.domain.user.entity.Generation;
