+++ src/test/java/com/devnity/devnity/domain/introduction/service/IntroductionServiceTest.java
@@ -15,6 +15,9 @@ import com.devnity.devnity.domain.introduction.dto.response.SuggestResponse;
 import com.devnity.devnity.domain.introduction.dto.response.UserIntroductionResponse;
 import com.devnity.devnity.domain.introduction.entity.Introduction;
 import com.devnity.devnity.domain.introduction.respository.IntroductionRepository;
+import com.devnity.devnity.web.introduction.service.IntroductionCommentService;
+import com.devnity.devnity.web.introduction.service.IntroductionLikeService;
+import com.devnity.devnity.web.introduction.service.IntroductionService;
 import com.devnity.devnity.web.user.dto.request.SaveIntroductionRequest;
 import com.devnity.devnity.domain.user.entity.Course;
 import com.devnity.devnity.domain.user.entity.Generation;
@@ -22,7 +25,7 @@ import com.devnity.devnity.domain.user.entity.Mbti;
 import com.devnity.devnity.domain.user.entity.User;
 import com.devnity.devnity.domain.user.entity.UserRole;
 import com.devnity.devnity.domain.user.repository.UserRepository;
-import com.devnity.devnity.web.user.service.UserRetrieveService;
+import com.devnity.devnity.domain.user.service.UserRetrieveService;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
