+++ src/test/java/com/devnity/devnity/domain/mapgakco/service/mapgakco/MapgakcoServiceTest.java
@@ -18,6 +18,8 @@ import com.devnity.devnity.domain.user.entity.User;
 import com.devnity.devnity.domain.user.entity.UserRole;
 import com.devnity.devnity.domain.user.repository.UserRepository;
 import java.time.LocalDateTime;
+
+import com.devnity.devnity.web.mapgakco.service.mapgakco.MapgakcoService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
