+++ src/test/java/com/devnity/devnity/domain/mapgakco/service/mapgakcoapplicant/MapgakcoApplicantServiceTest.java
@@ -17,6 +17,8 @@ import com.devnity.devnity.domain.user.entity.Generation;
 import com.devnity.devnity.domain.user.entity.User;
 import com.devnity.devnity.domain.user.entity.UserRole;
 import java.time.LocalDateTime;
+
+import com.devnity.devnity.web.mapgakco.service.mapgakcoapplicant.MapgakcoApplicantService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
